requirements.txt
streamlit
pandas 
matplotlib.pyplot
scipy.stats 
seaborn 